Add store class
