+++ public/index.html
@@ -17,7 +17,7 @@
       href="https://fonts.googleapis.com/css2?family=Zen+Loop&display=swap"
       rel="stylesheet"
     />
-    <script src="./dist/book.js" type="module"></script>
+    <script src="./dist/index.js" type="module"></script>
     <title>Awesome Books</title>
   </head>
   <body class="container text-center">
