+++ src/library.ts
@@ -1,0 +1,13 @@
+import type Book from "./types";
+
+export default class Library implements Book {
+  title;
+  author;
+  id;
+
+  constructor(title: string, author: string, id: string) {
+    this.title = title;
+    this.author = author;
+    this.id = id
+  }
+}
