+++ src/types.ts
@@ -2,7 +2,6 @@ type Book = {
   id: string;
   title: string;
   author: string;
-  
 }
 
 export default Book;
